FIX: Disable `/blog` URLs
