+++ public/markdown/blog-en-maintenance.md
@@ -1,0 +1,3 @@
+# En maintenance
+
+Notre blog est victime d'un incident de sécurité, l'accès en est désactivé temporairement. Notre équipe investigue pour un retour à la normale rapide. Veuillez nous excuser de ce désagrément.
